eni: Move limits to seperate package
[ upstream commit d69045af67c6ededf55bfc8c6252a02dd0411588 ]

This commit was necessary to prevent an import cycle in a future commit.
